Move to middleware folder in Edge.Hub.Http project and address comments.
